Добавил обработку алкогольных продуктов. Произвел разделение на пакеты
