+++ src/net/volgatech/CashDesk.java
@@ -1,6 +1,9 @@
 package net.volgatech;
 
-import java.lang.ref.Reference;
+import net.volgatech.Customer.Customer;
+import net.volgatech.Customer.PaymentMethod;
+import net.volgatech.Report.Report;
+import net.volgatech.Report.Bill;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
