--- src/net/volgatech/Handler.java
@@ -1,7 +1,0 @@
-package net.volgatech;
-
-public class Handler {
-    public void PickUp() {
-
-    }
-}
